# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/todo/type": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "查詢 TodoType 清單，支援關鍵字與排序",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TodoTypes"
                ],
                "summary": "取得 TodoType 列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "頁碼（預設 1）",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每頁筆數（預設 10）",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "關鍵字搜尋",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序欄位與方式，如 created_at desc",
                        "name": "order",
                        "in": "query"
                    }
                ],
                "responses": {}
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "建立一個新的 TodoType 項目",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TodoTypes"
                ],
                "summary": "新增 TodoType",
                "parameters": [
                    {
                        "description": "建立 TodoType 所需資料",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.TodoTypeCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "建立成功回傳的 TodoType 資料",
                        "schema": {
                            "$ref": "#/definitions/models.TodoTypes"
                        }
                    }
                }
            }
        },
        "/api/todo/type/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根據 ID 取得 TodoType 詳細資料",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TodoTypes"
                ],
                "summary": "取得單一 TodoType",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "TodoType ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功回傳 TodoType",
                        "schema": {
                            "$ref": "#/definitions/models.TodoTypes"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根據 ID 修改 TodoType 名稱",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TodoTypes"
                ],
                "summary": "修改 TodoType",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "TodoType ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "要更新的 TodoType 資料",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.TodoTypeUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功回傳更新後的 TodoType",
                        "schema": {
                            "$ref": "#/definitions/models.TodoTypes"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根據 ID 刪除指定的 TodoType",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TodoTypes"
                ],
                "summary": "刪除 TodoType",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "TodoType ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功回傳被刪除的 TodoType",
                        "schema": {
                            "$ref": "#/definitions/models.TodoTypes"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dto.TodoTypeCreateRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "dto.TodoTypeUpdateRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
...
        "models.TodoTypes": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "integer"
                },
                "deleted_by": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "integer"
                }
            }
        },
...
    }
}